uclf/requirements.txt
django
django-nose
facebook-sdk
social-auth-app-django
pyyaml
mysqlclient
django-mathfilters